requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dill==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
httpx==0.24.1
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.6.1
mypy-extensions==1.0.0
passlib==1.7.4
platformdirs==3.11.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pylint==3.0.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tomlkit==0.12.2
types-passlib==1.7.7.13
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
uvloop==0.19.0; platform_system != 'Windows'
